+++ CHANGELOG.rst
@@ -7,3 +7,4 @@ CHANGELOG
 - Add the /v1/ info page (#10)
 - Add the release status checks
 - Add the releasenotes bot (#16)
+- Add the archive.mozilla.org bot (#17)